refactor server setup into own dir
