+++ lib/middleware.js
@@ -5,7 +5,8 @@ const postgres = require("./postgres-db");
 const mongo = require("./mongo-db");
 const parseRequest = require("./parse-request");
 const extractPayload = require("./extract-payload");
-const { wss, WebSocket } = require("./server-config");
+const { wss } = require("../bin/www/server");
+const { WebSocket } = require("ws");
 
 const ENDPOINT_ID_LENGTH = 8;
 const REQUEST_ID_LENGTH = 16;
@@ -22,7 +23,6 @@ function handleFavicon(req, res, next) {
 }
 
 // app middleware
-/*make new endpoint then GET redirect to "/req/:endpointID"*/
 async function createNewEndpoint(_req, res) {
   const newEndpointID = cryptoStr(ENDPOINT_ID_LENGTH);
   const result = await postgres.postEndpoint(newEndpointID);
@@ -48,11 +48,9 @@ async function getRequestsHandler(req, res) {
     return res.status(400).send({ error: "Could not process request" });
   }
 
-  // req.session.endpoint = endpointHash;
   res.status(200).send(requests);
 }
 
-/*render the entire page with request content*/
 async function getPayloadHandler(req, res) {
   const { requestHash } = req.params;
   const request = await postgres.findRequest(requestHash);
