+++ lib/parse-request.js
@@ -1,11 +1,12 @@
 function parseRequest(req) {
-  const { method, path } = req;
-  const endpointHash = req.params.endpointHash
+  const { method, baseUrl, path } = req;
+  const endpointHash = req.params.endpointHash;
+  const fullPath = baseUrl + path;
 
   return {
     endpointHash,
     method,
-    path,
+    fullPath,
   };
 }
